+++ source/library/Language/Haskell/Brittany/Internal/Layouters/Import.hs
@@ -3,7 +3,6 @@
 module Language.Haskell.Brittany.Internal.Layouters.Import where
 
 import qualified Data.Semigroup as Semigroup
-import qualified Data.Set as Set
 import qualified Data.Text as Text
 import GHC (GenLocated(L), Located, moduleNameString, unLoc)
 import GHC.Hs
@@ -14,7 +13,6 @@ import GHC.Types.SourceText (SourceText(..), StringLiteral(..))
 import GHC.Unit.Types (IsBootInterface(..))
 import Language.Haskell.Syntax.ImpExp (ImportListInterpretation(EverythingBut))
 import Language.Haskell.Brittany.Internal.Config.Types
-import Language.Haskell.Brittany.Internal.ExactPrintCompat (AnnKey)
 import Language.Haskell.Brittany.Internal.Fallbacks (FallbackId(..))
 import Language.Haskell.Brittany.Internal.LayouterBasics
 import Language.Haskell.Brittany.Internal.Layouters.IE (layoutLLIEs, layoutAnnAndSepLLIEs, toL, SortItemsFlag(ShouldSortItems))
@@ -41,16 +39,12 @@ layoutImport :: ToBriDoc ImportDecl
 layoutImport = layoutImportWithExactText Nothing
 
 layoutImportWithExactText
-  :: Maybe (Text.Text, Set.Set AnnKey)
+  :: Maybe ExactSourceFragment
   -> Located (ImportDecl GhcPs)
   -> ToBriDocM BriDocNumbered
 layoutImportWithExactText exactText importNode@(L _ importD) = case exactText of
-  Just (source, annotationKeys) ->
-    briDocByExactTextWithAnnsNoComment
-      ImportFallback
-      importNode
-      annotationKeys
-      source
+  Just fragment -> briDocByExactSourceFragmentNoComment
+    ImportFallback importNode fragment
   Nothing -> layoutImportNormally importD
 
 layoutImportNormally :: ImportDecl GhcPs -> ToBriDocM BriDocNumbered